restructured project's tests
