--- src/tests/authconfig.json
@@ -1,3 +1,0 @@
-{
-	"https://simba-dev-api_platform_simbachain_com/v2/": {}
-}
