+++ tests/integration/test_deploy.ts
@@ -1,7 +1,7 @@
 import {
     SimbaConfig,
 } from "@simbachain/web3-suites";
-import {deployContract} from "../../commands/deploy";
+import {deployContract} from "../../src/commands/deploy";
 import { expect } from 'chai';
 import {default as chalk} from 'chalk';
 import axios from "axios";
@@ -11,7 +11,7 @@ const deployInfo = {
     url: "v2/organisations/20e69814-43d0-42b4-8499-d13a9d1afb23/contract_designs/6bbd0b0d-c5b2-4488-90a9-8357b77f1850/deploy/",
     blockchain: "Quorum",
     storage: "azure",
-    api: "ourtestapi",
+    api: "ourtestapi4",
     args: {
         _ourNum: 13,
         _ourString: "testing",
@@ -27,6 +27,7 @@ describe('tests deploy', () => {
 
         let detail: any;
         const res = await deployContract(undefined, deployInfo) as any;
+        console.log("resss: ", JSON.stringify(res))
         if (axios.isAxiosError(res) && res.response) {
             const data = res.response.data as any;
             detail = data.errors[0].detail;
