+++ tsconfig.json
@@ -23,6 +23,7 @@
     "./node_modules"
   ],
   "include": [
-    "./src"
+    "./src",
+    "tests"
   ]
 }
